refactor/KAN-16 changeg data types in AuthForm, made utils file for register page tests
